+++ HISTORY.md
@@ -1,5 +1,10 @@
 # Revision History for "py_dft"
 
+## Revision 0.4.0
+- Added code to implement the full solution to the schrodiger
+  equation. (The function _getPsi in schrodinger.py still has
+  bugs in it).
+
 ## Revision 0.3.0
 - Added the first three 4 sectios of assignment 2 to the
   schrodinger.py script (_getE, _H, _getgrad, _diagprod, and